Huge remodeling of the way gallery search, album and tag filtering works.
Every method of filtering tags or albums, searching was throwing SQL logic errors. I am bad at SQL so I came up with alternative methods of getting the pictures which work better. I think the issue is the SQL methods don't work well when 1 tag applies to many pictures in this context. The remodeling is ugly in a couple of places, and hopefully we can come up with simple LINQ sql queries but for now this works much better.